requirements.txt
langchain-openai
python-dotenv
PyGithub
fastapi
uvicorn